# 271 recursive simplification of constraints during evaluation