--- a/library/Booster/JsonRpc.hs
+++ b/library/Booster/JsonRpc.hs
@@ -177,22 +177,13 @@ respond stateVar =
                         (Right newPattern, patternTraces) -> do
                             let (term, mbPredicate, mbSubstitution) = externalisePattern newPattern
                                 tSort = externaliseSort (sortOfPattern newPattern)
-                                predicate = fromMaybe (KoreJson.KJTop tSort) mbPredicate
-                                substitution = fromMaybe (KoreJson.KJTop tSort) mbSubstitution
-                                result = KoreJson.KJAnd tSort [term, predicate, substitution]
-                            -- pure . Right . RpcTypes.Simplify $
-                            --     RpcTypes.SimplifyResult
-                            --         { state = addHeader result
-                            --         , logs = mkTraces patternTraces
-                            --         }
+                                result =
+                                    case catMaybes [mbPredicate, mbSubstitution] of
+                                        [] -> term
+                                        xs -> KoreJson.KJAnd tSort (term : xs)
                             pure $ Right (addHeader result, patternTraces)
                         (Left ApplyEquations.SideConditionsFalse{}, patternTraces) -> do
                             let tSort = fromMaybe (error "unknown sort") $ sortOfJson req.state.term
-                            -- pure . Right . RpcTypes.Simplify $
-                            --     RpcTypes.SimplifyResult
-                            --         { state = addHeader $ KoreJson.KJBottom tSort
-                            --         , logs = mkTraces patternTraces
-                            --         }
                             pure $ Right (addHeader $ KoreJson.KJBottom tSort, patternTraces)
                         (Left (ApplyEquations.EquationLoop terms), _traces) ->
                             pure . Left . RpcError.backendError RpcError.Aborted $ map externaliseTerm terms -- FIXME
@@ -207,11 +198,6 @@ respond stateVar =
                                     fromMaybe (error "not a predicate") $
                                         sortOfJson req.state.term
                                 result = externalisePredicate predicateSort newPred
-                            -- pure . Right . RpcTypes.Simplify $
-                            --     RpcTypes.SimplifyResult
-                            --         { state = addHeader result
-                            --         , logs = mkTraces traces
-                            --         }
                             pure $ Right (addHeader result, traces)
                         (Left something, _traces) ->
                             pure . Left . RpcError.backendError RpcError.Aborted $ show something -- FIXME