--- a/test/rpc-integration/resources/simplify.k
+++ b/test/rpc-integration/resources/simplify.k
@@ -14,11 +14,11 @@ module SIMPLIFY
   // an indirect simplifiation loop
   rule [p1-if-p2]: p1(X) => true requires p2(X) [simplification]
   rule [p2-if-p1]: p2(X) => true requires p1(X) [simplification]
-  // simplification of terms involving p1(_someInt) or p2(_someInt) will
-  // enter a loop of p1, p2, p1, p2... until depth 100 is reached, and then
-  // bail out in booster. Simplification in kore-rpc _loops to infinity_.
+  // simplification of terms involving p1(_someInt) or p2(_someInt) will enter
+  // a loop of p1, p2, p1, p2... until the maximum recursion depth is reached
+  // and then bail out in booster. Simplification in kore-rpc _loops to infinity_.
 
-  // a set of fucntion equations that require second-order evaluation of side condition p3
+  // a set of function equations that require recursive evaluation of side condition p3
   rule [eval-g]: g(X) => X requires p3(X)
 
   rule [p3-true]: p3(1) => true