--- a/test/rpc-integration/runDirectoryTest.sh
+++ b/test/rpc-integration/runDirectoryTest.sh
@@ -100,10 +100,10 @@ elif [ -d $dir ]; then
             params=""
         fi
         # call rpc-client
-        if [ $(basename $server) == "booster-dev" ] && [ -f "$dir/response-${testname}.booster-dev" ]; then
+        if [ "$(basename ${server[0]})" == "booster-dev" ] && [ -f "$dir/response-${testname}.booster-dev" ]; then
             echo "$client $mode $test $params --expect $dir/response-${testname}.booster-dev $*"
             $client $mode $test $params --expect $dir/response-${testname}.booster-dev $*
-        elif [ $(basename $server) == "kore-rpc-dev" ] && [ -f "$dir/response-${testname}.kore-rpc-dev" ]; then
+        elif [ "$(basename ${server[0]})" == "kore-rpc-dev" ] && [ -f "$dir/response-${testname}.kore-rpc-dev" ]; then
             echo "$client $mode $test $params --expect $dir/response-${testname}.kore-rpc-dev $*"
             $client $mode $test $params --expect $dir/response-${testname}.kore-rpc-dev $*
         else