--- a/test/rpc-integration/test-simplify/state-evaluate-two-stage.simplify
+++ b/test/rpc-integration/test-simplify/state-evaluate-two-stage.simplify
@@ -1,0 +1,20 @@
+{
+  "format": "KORE",
+  "version": 1,
+  "term": {
+                "tag": "App",
+                "name": "Lblg",
+                "sorts": [],
+                "args": [
+                  {
+                    "tag": "DV",
+                    "sort": {
+                      "tag": "SortApp",
+                      "name": "SortInt",
+                      "args": []
+                    },
+                    "value": "1"
+                  }
+                ]
+              }
+}