--- a/test/rpc-integration/test-simplify/state-evaluate-under-function.simplify
+++ b/test/rpc-integration/test-simplify/state-evaluate-under-function.simplify
@@ -43,7 +43,7 @@
                   "name": "SortInt",
                   "args": []
                 },
-                "value": "34"
+                "value": "0"
               }
             ]
           }