--- a/test/rpc-integration/test-simplify/state-simplification-loop.simplify-do-not-use
+++ b/test/rpc-integration/test-simplify/state-simplification-loop.simplify-do-not-use
@@ -1,0 +1,20 @@
+{
+  "format": "KORE",
+  "version": 1,
+  "term": {
+                "tag": "App",
+                "name": "Lblp1",
+                "sorts": [],
+                "args": [
+                  {
+                    "tag": "DV",
+                    "sort": {
+                      "tag": "SortApp",
+                      "name": "SortInt",
+                      "args": []
+                    },
+                    "value": "42"
+                  }
+                ]
+              }
+}