… test/rpc-integration/resources/simplify.kore
… test/rpc-integration/test-simplify/README.md
… test/rpc-integration/test-simplify/response-evaluate-two-stage-fail.json
… test/rpc-integration/test-simplify/response-evaluate-two-stage.json
… test/rpc-integration/test-simplify/response-evaluate-under-function.json
… test/rpc-integration/test-simplify/response-if-then-else-arity-error.json
… test/rpc-integration/test-simplify/response-if-then-else-sort-error.json
… test/rpc-integration/test-simplify/response-not-and-false1.json
… test/rpc-integration/test-simplify/response-not-and-false2.json
… test/rpc-integration/test-simplify/response-simplification-loop.json
… test/rpc-integration/test-simplify/response-symbolic-first-of-3.json
… test/rpc-integration/test-simplify/response-with-logging.json
… test/rpc-integration/test-simplify/state-if-then-else-eval.simplify
… test/rpc-integration/test-simplify/state-not-and-false1.simplify
… test/rpc-integration/test-simplify/state-not-and-false2.simplify
… test/rpc-integration/test-simplify/state-simplification-loop.simplify