* recursive simplification of side conditions in rewrite rules and equations
* restrict recursion depth to a small value (5, not 100)
* modified simplification integration tests, including a description.  
   - now executes with booster-dev only
  - added a test that would loop in kore-rpc-booster or kore-rpc-dev

Fixes #271