@jberthold 2024-01-12
KEVM timings vs latest master: a few improvements

| Test                              | 271-recursive-simplification-reprise time | master-a673bb5 time | (271-recursive-simplification-reprise/master-a673bb5) time 
|-----------------------------------|-------------------------------------------|---------------------|------------------------------------------------------------
| benchmarks/encode-keccak00-spec.k | 43.47                                     | 48.37               | 0.8986975397973951                                         
| benchmarks/bytes00-spec.k         | 67.29                                     | 71.64               | 0.9392797319933                                            
| erc20/ds/approve-success-spec.k   | 55.59                                     | 58.38               | 0.9522096608427544                                         
| erc20/hkg/approve-spec.k          | 46.46                                     | 48.13               | 0.9653023062538957                                         
| TOTAL                             | 212.81000000000003                        | 226.51999999999998  | 0.939475542998411