@geo2a 2024-01-12
Kontrol performance (5 workers): some improvements and some really weird looking regressions, perhaps due to noise? rerunning with 1 worker.
| Test                                                                   | 271-recursive-simplification-reprise time | master-a673bb5 time | (271-recursive-simplification-reprise/master-a673bb5) time
|------------------------------------------------------------------------|-------------------------------------------|---------------------|------------------------------------------------------------
| AssumeTest.test_multi_assume(address,address)                          | 143.23                                    | 197.72              | 0.7244082540967024
| StoreTest.testGasLoadWarmUp()                                          | 75.22                                     | 98.43               | 0.7641979071421314
| ExpectRevertTest.test_expectRevert_message()                           | 113.35                                    | 132.59              | 0.8548910174221283
| SymbolicStorageTest.testEmptyInitialStorage(uint256)                   | 50.98                                     | 59.19               | 0.8612941375232303
| SymbolicStorageTest.testFail_SymbolicStorage(uint256)                  | 141.08                                    | 161.01              | 0.8762188683932677
| AllowChangesTest.testFailAllowChangesToStorage()                       | 197.52                                    | 217.94              | 0.9063044874736167
| BytesTypeTest.test_bytes4(bytes4)                                      | 48.9                                      | 51.63               | 0.94712376525276
| ArithmeticTest.test_max1(uint256,uint256)                              | 103.38                                    | 107.1               | 0.965266106442577
| StoreTest.testStoreLoadNonExistent()                                   | 99.69                                     | 95.64               | 1.0423462986198244
| LabelTest.testLabel()                                                  | 37.89                                     | 36.04               | 1.0513318534961156
| AddrTest.test_builtInAddresses()                                       | 28.35                                     | 26.67               | 1.0629921259842519
| ExpectRevertTest.test_ExpectRevert_increasedDepth()                    | 204.28                                    | 188.54              | 1.0834836109048478
| BlockParamsTest.testWarp(uint256)                                      | 46.82                                     | 43.04               | 1.087825278810409
| BlockParamsTest.testChainId(uint256)                                   | 54.55                                     | 44.42               | 1.2280504277352542
| StoreTest.testGasStoreWarmUp()                                         | 98.69                                     | 73.2                | 1.3482240437158468
| ArithmeticTest.test_max2(uint256,uint256)                              | 89.86                                     | 65.96               | 1.3623408126137053
| src/tests/integration/test_foundry_prove.py::test_foundry_resume_proof | 117.51                                    | 86.13               | 1.3643329850226402
| AssumeTest.test_assume_false(uint256,uint256)                          | 111.81                                    | 80.91               | 1.381905821282907
| StoreTest.testLoadNonExistent()                                        | 66.73                                     | 46.22               | 1.443747295543055
| BlockParamsTest.testFee(uint256)                                       | 62.07                                     | 42.6                | 1.4570422535211267
| BlockParamsTest.testBlockNumber()                                      | 39.45                                     | 25.76               | 1.531444099378882
| BlockParamsTest.testCoinBase()                                         | 58.11                                     | 37.53               | 1.5483613109512389
| BlockParamsTest.testRoll(uint256)                                      | 64.72                                     | 40.51               | 1.5976302147617873
| BMCLoopsTest.test_countdown_concrete()                                 | 25.32                                     | 4.52                | 5.601769911504426
| TOTAL                                                                  | 2079.5099999999998                        | 1963.3000000000004  | 1.0591911577446134