@ehildenb 2024-01-12
Do we have any examples where this allows execution to continue where it couldn't before? If we don't see significant performance improvements, then we should at least check that we are indeed falling back less. Can we post fallback statistics here?