@jberthold 2024-01-14
> Do we have any examples where this allows execution to continue where it couldn't before? If we don't see significant performance improvements, then we should at least check that we are indeed falling back less. Can we post fallback statistics here?

The code change enables jump condition simplifications when combined with the `evm-semantics` change to add `preserves-definedness` attributes to some key simplifications.
 
> The worry I have is that recursive constraint simplification has _always_ caused problems in the past and been tricky to implement, usually with creating potentially infinite looping code. Do we have any protections against infinite looping? Is the depth-limit of 5 mentioned above doing that? Is that user-configurable?

The limit of 5 is currently not configurable but we can add an option to adjust that limit. There is one test with two mutually-recursive simplifications, which indeed loops forever in the legacy backend but is caught by the new code.