@geo2a 2024-01-15
Kontrol performance with 3 workers
| Test                                                                   | 271-recursive-simplification-reprise time | master-a673bb5 time | (271-recursive-simplification-reprise/master-a673bb5) time
|------------------------------------------------------------------------|-------------------------------------------|---------------------|------------------------------------------------------------
| Setup2Test.test_setup()                                                | 42.73                                     | 57.1                | 0.7483362521891418
| BytesTypeTest.test_bytes4(bytes4)                                      | 29.17                                     | 33.64               | 0.8671224732461356
| AllowChangesTest.testFailAllowChangesToStorage()                       | 183.9                                     | 196.82              | 0.9343562646072554
| ArithmeticTest.test_max1(uint256,uint256)                              | 96.85                                     | 93.47               | 1.0361613351877608
| BMCLoopsTest.test_countdown_concrete()                                 | 25.04                                     | 24.04               | 1.0415973377703827
| AccountParamsTest.test_Nonce_NonExistentAddress()                      | 63.54                                     | 60.89               | 1.0435211036294958
| BlockParamsTest.testWarp(uint256)                                      | 42.04                                     | 40.25               | 1.044472049689441
| AssumeTest.test_assume_staticCall(bool)                                | 54.68                                     | 52.09               | 1.0497216356306391
| Empty.sol                                                              | 68.16                                     | 64.72               | 1.0531520395550062
| MethodDisambiguateTest.test_method_call()                              | 26.11                                     | 23.95               | 1.090187891440501
| ArithmeticTest.test_max2(uint256,uint256)                              | 80.1                                      | 61.82               | 1.2956971853769006
| StoreTest.testGasLoadWarmUp()                                          | 92.41                                     | 71.05               | 1.3006333567909922
| src/tests/integration/test_foundry_prove.py::test_foundry_resume_proof | 111.21                                    | 80.36               | 1.3838974614235937
| AssumeTest.test_assume_false(uint256,uint256)                          | 105.81                                    | 75.11               | 1.4087338570097192
| AssumeTest.testFail_assume_true(uint256,uint256)                       | 99.32                                     | 68.66               | 1.4465482085639383
| TOTAL                                                                  | 1121.07                                   | 1003.97             | 1.1166369513033256